airflow/requirements.txt
apache-airflow-providers-google
pyarrow
pandas
sqlalchemy
psycopg2-binary
# dbt-bigquery==1.0.0